# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Symbol {
  id: number;
  symbol: string;
  name: string;
  addedAt: string;
}

export interface SymbolInput {
  /** @minLength 1 */
  symbol: string;
  /** @minLength 1 */
  name: string;
}

export type SignalSide = typeof SignalSide[keyof typeof SignalSide];


export const SignalSide = {
  long: 'long',
  short: 'short',
} as const;

export type SignalRiskTag = typeof SignalRiskTag[keyof typeof SignalRiskTag];


export const SignalRiskTag = {
  Safe: 'Safe',
  Medium: 'Medium',
  Danger: 'Danger',
} as const;

export type SignalState = typeof SignalState[keyof typeof SignalState];


export const SignalState = {
  active: 'active',
  tp_hit: 'tp_hit',
  sl_hit: 'sl_hit',
  expired: 'expired',
} as const;

export interface Signal {
  id: number;
  signalId: string;
  symbol: string;
  timeframe: string;
  barTime: string;
  side: SignalSide;
  entryPrice: number;
  slPrice: number;
  tpPrice: number;
  /** @nullable */
  currentSlPrice?: number | null;
  confidence: number;
  riskTag: SignalRiskTag;
  state: SignalState;
  /** @nullable */
  exitPrice?: number | null;
  /** @nullable */
  exitReason?: string | null;
  /** @nullable */
  exitBarTime?: string | null;
  /** @nullable */
  rrRatio?: number | null;
  /** @nullable */
  pattern?: string | null;
  /** @nullable */
  regime?: string | null;
  createdAt: string;
}

export interface SignalStats {
  total: number;
  active: number;
  tp_hit: number;
  sl_hit: number;
  expired: number;
  winRate: number;
  avgConfidence: number;
  avgRR: number;
}

export interface Bar {
  /** Unix timestamp in seconds */
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export type ListSignalsParams = {
symbol?: string;
limit?: number;
};

export type GetSignalStatsParams = {
symbol?: string;
};

export type ListBarsParams = {
symbol: string;
timeframe?: string;
limit?: number;
};