# lib/api-zod/src/generated/types/bar.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Bar {
  /** Unix timestamp in seconds */
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}