# lib/api-zod/src/generated/types/signal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SignalRiskTag } from './signalRiskTag';
import type { SignalSide } from './signalSide';
import type { SignalState } from './signalState';

export interface Signal {
  id: number;
  signalId: string;
  symbol: string;
  timeframe: string;
  barTime: Date;
  side: SignalSide;
  entryPrice: number;
  slPrice: number;
  tpPrice: number;
  /** @nullable */
  currentSlPrice?: number | null;
  confidence: number;
  riskTag: SignalRiskTag;
  state: SignalState;
  /** @nullable */
  exitPrice?: number | null;
  /** @nullable */
  exitReason?: string | null;
  /** @nullable */
  exitBarTime?: Date | null;
  /** @nullable */
  rrRatio?: number | null;
  /** @nullable */
  pattern?: string | null;
  /** @nullable */
  regime?: string | null;
  createdAt: Date;
}