# lib/api-zod/src/generated/types/signalStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SignalStats {
  total: number;
  active: number;
  tp_hit: number;
  sl_hit: number;
  expired: number;
  winRate: number;
  avgConfidence: number;
  avgRR: number;
}