# lib/api-zod/src/generated/types/symbol.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Symbol {
  id: number;
  symbol: string;
  name: string;
  addedAt: Date;
}