# lib/api-zod/src/generated/types/symbolInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SymbolInput {
  /** @minLength 1 */
  symbol: string;
  /** @minLength 1 */
  name: string;
}